- Fixes done on      PetService      EmployeeService      and CustomerService

+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/pet/PetController.java
@@ -32,6 +32,6 @@ public class PetController {
 
     @GetMapping("/owner/{ownerId}")
     public List<PetDTO> getPetsByOwner(@PathVariable long ownerId) {
-        return petService.getPetsByOwner(ownerId);
+        return petService.getPetsByOwner((long)ownerId);
     }
 }
